move TrainingAverager to optim
